Refactor into date-utils
