+++ src/app/feed/feed.page.ts
@@ -30,6 +30,7 @@ import { PlatformService } from '../services/platform.service';
 import { SettingsService } from '../services/settings.service';
 import { SourcesService } from '../services/sources.service';
 import { SettingsComponent } from '../settings/settings.component';
+import { formatDateAsDay, formatDateAsLong, formatDateRelative } from '../lib/date-utils';
 
 @Component({
   selector: 'app-feed',
@@ -57,51 +58,6 @@ export class FeedPage {
               public settingsService: SettingsService) {
     addIcons({ bookmark, bookmarkOutline, shareSocialOutline, ellipsisVertical, filterOutline, chevronForward, 
       chevronUpOutline });
-  }
-
-  public formatDate(dateStr: string | number, locale: string) {
-    const date = new Date(dateStr);
-    const today = new Date();
-    const yesterday = new Date();
-    yesterday.setDate(today.getDate() - 1);
-
-    // Remove time for comparison
-    const dateOnly = new Date(date.getFullYear(), date.getMonth(), date.getDate());
-    const todayOnly = new Date(today.getFullYear(), today.getMonth(), today.getDate());
-    const yesterdayOnly = new Date(yesterday.getFullYear(), yesterday.getMonth(), yesterday.getDate());
-
-    if (dateOnly.getTime() === todayOnly.getTime()) {
-      return 'Today';
-    } else if (dateOnly.getTime() === yesterdayOnly.getTime()) {
-      return 'Yesterday';
-    } else {
-      const dateOptions: Intl.DateTimeFormatOptions = {
-        year: 'numeric',
-        month: 'short',
-        day: 'numeric'
-      };
-      return date.toLocaleDateString(locale, dateOptions);
-    }
-  }
-
-  public formatDateAsDay(dateStr: string | number, locale: string) {
-    const date = new Date(dateStr);
-    const dateOptions: Intl.DateTimeFormatOptions = {
-      weekday: 'short'
-    };
-    return date.toLocaleDateString(locale, dateOptions);     
-  }
-
-  public formatDateAsLong(dateStr: string | number, locale: string) {
-    const date = new Date(dateStr);
-    const dateOptions: Intl.DateTimeFormatOptions = {
-      hour: 'numeric',
-      minute: 'numeric',
-      weekday: 'short',
-      month: 'short',
-      day: 'numeric'
-    };
-    return date.toLocaleDateString(locale, dateOptions);        
   }
 
   public scrollToTop() {
